heypico-be/requirements.txt
fastapi[standard]
python-dotenv
google-genai
# openai